+++ soridam-infra/src/main/java/sorisoop/soridam/infra/notification/NotificationAsyncService.java
@@ -10,6 +10,8 @@ import lombok.extern.slf4j.Slf4j;
 import sorisoop.soridam.domain.favoriteplace.application.FavoritePlaceQueryService;
 import sorisoop.soridam.domain.notification.application.NotificationCommandService;
 import sorisoop.soridam.domain.notification.domain.Notification;
+import sorisoop.soridam.infra.notification.sse.NotificationSsePayload;
+import sorisoop.soridam.infra.notification.sse.SseEmitterManager;
 
 @Slf4j
 @Component
@@ -21,25 +23,21 @@ public class NotificationAsyncService {
 
 	@Async
 	public void sendReviewNotification(ReviewCreatedEvent event) {
-		try {
-			List<Notification> notifications = favoritePlaceQueryService.findTargetsByPlaceIdExcludingUser(event.placeId(),
-					event.writerId()).stream()
-				.map(favoritePlace -> {
-					Long userId = favoritePlace.userId();
-					Long placeId = favoritePlace.placeId();
-					String placeName = favoritePlace.placeName();
-					String content = String.format("즐겨찾기한 장소 %s에 새로운 리뷰가 등록되었습니다.", placeName);
-					return Notification.createReviewNotification(userId, placeId, content);
-				})
-				.toList();
-			notificationCommandService.createNotifications(notifications);
+		List<Notification> notifications = favoritePlaceQueryService.findTargetsByPlaceIdExcludingUser(event.placeId(),
+				event.writerId()).stream()
+			.map(favoritePlace -> {
+				Long userId = favoritePlace.userId();
+				Long placeId = favoritePlace.placeId();
+				String placeName = favoritePlace.placeName();
+				String content = String.format("즐겨찾기한 장소 %s에 새로운 리뷰가 등록되었습니다.", placeName);
+				return Notification.createReviewNotification(userId, placeId, content);
+			})
+			.toList();
+		notificationCommandService.createNotifications(notifications);
 
-			notifications.forEach(notification -> {
-				NotificationSsePayload payload = NotificationSsePayload.from(notification);
-				sseEmitterManager.sendToUser(notification.getReceiverId(), payload);
-			});
-		} catch (Exception e) {
-			log.error("리뷰 알림 전송 중 오류 발생: {}", e.getMessage(), e);
-		}
+		notifications.forEach(notification -> {
+			NotificationSsePayload payload = NotificationSsePayload.from(notification);
+			sseEmitterManager.sendToUser(notification.getReceiverId(), payload);
+		});
 	}
 }
